pyproject.toml
[project]
name = "assignment-05-secure-data-encryption-system-using-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
